+++ HISTORY.rst
@@ -6,8 +6,9 @@ Changelog
 0.2.0 (unreleased)
 ==================
 
-- Add `--web` option to `search` command.
+- Add ``--web`` option to `search` command.
 - Fix bug with locating packages with an underscore character.
+- Add license, maintainer, and documentation info to the ``info`` command.
 
 
 0.1.0 (2014-06-08)